Refactoring UI out of HomeViewController
